use RESTFul style for url、new primary key style.
(1) site back-end
1、modify IRequest as service input、IResponse as service result.
2、move all ajax request to alone api controller.
3、use new id generator：SnowflakeIdWorker.
4、use new primary key style, the front end uses a short id, the back end uses a long id, and automatically converts.
5、update old primary key style to new.
6、modify url to RESTFul style.
7、add auto convert annotation: URLPrefixFill(file path auto fill with cdn full address)、GeneralConvert、PrimaryKeyConvert.
8、add access interceptor to automatically record access to user creations.
9、unified exclusion filter with ExcludedFilter.

（2）site front-end
1、reorganize css of modal and profilecenter and so forth.
2、reorganize css and js of edit page.
3、show a default img when img load error.

（3）user module
1、user's feed flow according to user's followings and the allow user groups by manager.
2、add history page for showing user's visited and liked creations.

（4）album module:
1、support define topic for photos.
2、support comment for tags and topic.
3、support like photo for user.
4、support more permission type.
5、add users's photo-tag-wrapper manager page.
6、support set auto mount by tagWrapper.name or photo.tags for album.

（5）video module
1、add upload mp3 file support.
2、support upload subtitle for video.
3、support topic comment for video.
4、support like video for user.
5、support more permission type.
6、add video player rotate、flip、loop control btn.
7、video iframe web fullscreen and auto setting by iframe attr.

（6）message service
1、support notify history unread message when user login.
2、support upload image in comment and letter.
3、support like comment.

+++ src/main/java/site/imcoder/blog/cache/FlushCacheTool.java
@@ -32,7 +32,7 @@ public class FlushCacheTool {
      * @param articleBuffer
      * @return int 成功与否
      */
-    public void flushArticleCache(Map<Integer, Article> articleBuffer, Set<Integer> hasUpdateArticle) {
+    public void flushArticleCache(Map<Long, Article> articleBuffer, Set<Long> hasUpdateArticle) {
 
         if (hasUpdateArticle.size() == 0) {
             logger.debug("FlushArticleCache: hasUpdateArticle 的大小为 0 ，跳过flush article");
@@ -43,24 +43,22 @@ public class FlushCacheTool {
 
         logger.info("FlushArticleCache: Cache触发执行flush, flush article数量：" + hasUpdateArticle.size());
 
-        for (int aid : hasUpdateArticle) {
+        for (Long aid : hasUpdateArticle) {
             Article article = articleBuffer.get(aid);
             if (article != null) {
                 articleList.add(article);
             }
         }
-        //重置hasUpdateArticle
+        // 重置hasUpdateArticle
         if (statsDao.saveArticleBuffer(articleList) > 0) {
-
             logger.debug("FlushArticleCache: 重置 hasUpdateArticle");
             logger.debug("FlushArticleCache: flush success!");
-
             hasUpdateArticle.clear();
         } else {
             String tips = "";
             if (hasUpdateArticle.size() > 0) {
                 tips = " 可能是文章 ";
-                for (int aid : hasUpdateArticle) {
+                for (Long aid : hasUpdateArticle) {
                     tips += (aid + ",");
                 }
                 tips = tips.substring(0, tips.length() - 1);
